Improve logging system, add Test Cases
- Change logging system convention.
- Add logging detail for tracing and investigating.
- Resolve backend bugs (input float data and saved float data have different value).

+++ go_mock/database/album.go
@@ -8,10 +8,10 @@ type Album struct {
 	ID     int64
 	Title  string
 	Artist string
-	Price  float32
+	Price  float64
 }
 
-func NewAlbulm(title, artist string, price float32) Album {
+func NewAlbulm(title, artist string, price float64) Album {
 	return Album{
 		Title:  title,
 		Artist: artist,
